pyproject.toml
[tool.poetry]
name = "hooks-toolkit"
version = "1.0.5"
description = ""
authors = ["Denis Angell <dangell@transia.co>"]

[tool.poetry.dependencies]
python = "^3.8"
python-dotenv = "^1.0.0"
xahau-py = "^3.0.0"

[tool.poetry.dev-dependencies]
pytest = "^7.4.0"
flake8 = "^3.8.4"